+++ rabix-backend-local/src/main/java/org/rabix/backend/local/CommandLinePrinter.java
@@ -4,22 +4,13 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.Set;
 
-import org.rabix.bindings.helper.URIHelper;
-import org.rabix.bindings.model.Context;
 import org.rabix.bindings.model.Job;
-import org.rabix.bindings.model.Job.JobStatus;
-import org.rabix.bindings.model.dag.DAGLinkPort.LinkPortType;
-import org.rabix.bindings.model.dag.DAGNode;
 import org.rabix.common.helper.FileHelper;
-import org.rabix.common.helper.InternalSchemaHelper;
-import org.rabix.common.json.BeanSerializer;
+import org.rabix.engine.JobHelper;
 import org.rabix.engine.db.DAGNodeDB;
-import org.rabix.engine.model.ContextRecord;
 import org.rabix.engine.model.JobRecord;
 import org.rabix.engine.model.JobRecord.PortCounter;
 import org.rabix.engine.model.LinkRecord;
@@ -72,7 +63,7 @@ public class CommandLinePrinter implements IterationCallback {
     if (!iterationDirectory.exists()) {
       iterationDirectory.mkdirs();
     }
-    List<Job> jobs = createJobs(eventProcessor, contextId);
+    Set<Job> jobs = JobHelper.createReadyJobs(jobRecordService, variableRecordService, contextRecordService, dagNodeDB, contextId);
     if (!jobs.isEmpty()) {
       File readyDirectory = new File(iterationDirectory, READY_JOBS_DIR);
       readyDirectory.mkdirs();
@@ -201,28 +192,6 @@ public class CommandLinePrinter implements IterationCallback {
     } catch (IOException e) {
       e.printStackTrace();
     }
-  }
-
-  private List<Job> createJobs(EventProcessor eventProcessor, String contextId) {
-    List<Job> jobs = new ArrayList<>();
-    List<JobRecord> jobRecords = jobRecordService.findReady(contextId);
-
-    if (!jobRecords.isEmpty()) {
-      for (JobRecord job : jobRecords) {
-        DAGNode node = dagNodeDB.get(InternalSchemaHelper.normalizeId(job.getId()), contextId);
-        Map<String, Object> inputs = new HashMap<>();
-
-        List<VariableRecord> inputVariables = variableRecordService.find(job.getId(), LinkPortType.INPUT, contextId);
-        for (VariableRecord inputVariable : inputVariables) {
-          inputs.put(inputVariable.getPortId(), inputVariable.getValue());
-        }
-        ContextRecord contextRecord = contextRecordService.find(job.getContextId());
-        Context context = new Context(contextRecord.getId(), contextRecord.getConfig());
-        String encodedApp = URIHelper.createDataURI(BeanSerializer.serializeFull(node.getApp()));
-        jobs.add(new Job(job.getExternalId(), job.getId(), encodedApp, JobStatus.READY, inputs, null, context));
-      }
-    }
-    return jobs;
   }
 
 }
