+++ rabix-backend-local/src/main/java/org/rabix/backend/local/LocalJobHandler.java
@@ -1,9 +1,6 @@
 package org.rabix.backend.local;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.Executors;
@@ -12,19 +9,11 @@ import org.rabix.bindings.BindingException;
 import org.rabix.bindings.Bindings;
 import org.rabix.bindings.BindingsFactory;
 import org.rabix.bindings.ProtocolType;
-import org.rabix.bindings.helper.URIHelper;
-import org.rabix.bindings.model.Context;
 import org.rabix.bindings.model.Job;
 import org.rabix.bindings.model.Job.JobStatus;
-import org.rabix.bindings.model.dag.DAGLinkPort.LinkPortType;
-import org.rabix.bindings.model.dag.DAGNode;
-import org.rabix.common.helper.InternalSchemaHelper;
-import org.rabix.common.json.BeanSerializer;
+import org.rabix.engine.JobHelper;
 import org.rabix.engine.db.DAGNodeDB;
 import org.rabix.engine.event.impl.JobStatusEvent;
-import org.rabix.engine.model.ContextRecord;
-import org.rabix.engine.model.JobRecord;
-import org.rabix.engine.model.VariableRecord;
 import org.rabix.engine.processor.EventProcessor;
 import org.rabix.engine.processor.EventProcessor.IterationCallback;
 import org.rabix.engine.service.ContextRecordService;
@@ -62,7 +51,7 @@ public class LocalJobHandler implements IterationCallback {
 
   @Override
   public void call(final EventProcessor eventProcessor, final String contextId, int iteration) {
-    List<Job> jobs = createJobs(eventProcessor, contextId);
+    Set<Job> jobs = JobHelper.createReadyJobs(jobRecordService, variableRecordService, contextRecordService, dagNodeDB, contextId);
     for (final Job job : jobs) {
       if (!runningJobs.contains(job)) {
         runningJobs.add(job);
@@ -101,30 +90,5 @@ public class LocalJobHandler implements IterationCallback {
         });
       }
     }
-  }
-
-  /**
-   * Creates ready Jobs 
-   */
-  private List<Job> createJobs(EventProcessor eventProcessor, String contextId) {
-    List<Job> jobs = new ArrayList<>();
-    List<JobRecord> jobRecords = jobRecordService.findReady(contextId);
-
-    if (!jobRecords.isEmpty()) {
-      for (JobRecord job : jobRecords) {
-        DAGNode node = dagNodeDB.get(InternalSchemaHelper.normalizeId(job.getId()), contextId);
-
-        Map<String, Object> inputs = new HashMap<>();
-        List<VariableRecord> inputVariables = variableRecordService.find(job.getId(), LinkPortType.INPUT, contextId);
-        for (VariableRecord inputVariable : inputVariables) {
-          inputs.put(inputVariable.getPortId(), inputVariable.getValue());
-        }
-        ContextRecord contextRecord = contextRecordService.find(job.getContextId());
-        Context context = new Context(contextRecord.getId(), contextRecord.getConfig());
-        String encodedApp = URIHelper.createDataURI(BeanSerializer.serializeFull(node.getApp()));
-        jobs.add(new Job(job.getExternalId(), job.getId(), encodedApp, JobStatus.READY, inputs, null, context));
-      }
-    }
-    return jobs;
   }
 }
