+++ rabix-engine-rest/src/main/java/org/rabix/engine/rest/service/JobService.java
@@ -1,6 +1,6 @@
 package org.rabix.engine.rest.service;
 
-import java.util.List;
+import java.util.Set;
 
 import org.rabix.bindings.model.Job;
 import org.rabix.engine.processor.EventProcessor;
@@ -9,6 +9,6 @@ public interface JobService {
 
   void update(Job job) throws JobServiceException;
   
-  List<Job> getReady(EventProcessor eventProcessor, String contextId) throws JobServiceException;
+  Set<Job> getReady(EventProcessor eventProcessor, String contextId) throws JobServiceException;
   
 }
